split utils so that external imports are allowed
